requirements.txt
numpy>=1.26.0
faiss-cpu>=1.7.4
sentence-transformers>=2.6.1
langchain-core>=0.1.10
langchain-openai>=0.1.3
langchain-community>=0.0.34
fastapi>=0.110.0
pydantic>=2.7.0
python-dotenv>=1.0.1
uvicorn>=0.27.0